requirements.txt
aiogram==3.1.1
asyncpg==0.29.0
python-dotenv==1.0.0
Requests==2.31.0
tabulate==0.9.0